night before

+++ src/main/java/team3176/robot/subsystems/superstructure/intake/IntakeIO.java
@@ -29,7 +29,7 @@ public interface IntakeIO {
     public boolean isPivotLinebreak = false;
     public boolean upperLimitSwitch = false;
     public boolean lowerLimitSwitch = false;
-    public int laserCanMeasurement = 0;
+    public double laserCanMeasurement = 0;
     public boolean isNotePresent = false;
 
     // constructor if needed for some inputs
